refactor: Simplify getter method resolution and update codec action implementations
