+++ src/main/kotlin/io/github/tt432/k9tools/GenerateAnvilLibCodecAction.kt
@@ -14,7 +14,7 @@ import com.intellij.psi.PsiTypeElement
 import com.intellij.psi.codeStyle.JavaCodeStyleManager
 
 /**
- * @author TT432
+ * @author QiuShui1012
  */
 @Suppress("ConstPropertyName", "DuplicatedCode")
 class GenerateAnvilLibCodecAction : AnAction() {
@@ -103,7 +103,7 @@ class GenerateAnvilLibCodecAction : AnAction() {
                     ${getCodecRef(it.typeElement)}
                     .${getFieldOf(it)}
                     .forGetter(${
-                        getGetterName(
+                        getGetter(
                             className,
                             it,
                             getFieldAndGetterMethod(psiClass)
