+++ src/main/kotlin/io/github/tt432/k9tools/GetterFinder.kt
@@ -1,0 +1,82 @@
+package io.github.tt432.k9tools
+
+import com.intellij.psi.PsiArrayInitializerMemberValue
+import com.intellij.psi.PsiClass
+import com.intellij.psi.PsiField
+import com.intellij.psi.PsiLiteralExpression
+
+/**
+ * @author QiuShui1012 & TT432
+ */
+
+fun getFieldAndGetterMethod(psiClass: PsiClass): Map<PsiField, String?> {
+    val isRecord = psiClass.isRecord
+    return psiClass.allFields.associateWith { field ->
+        if (isRecord) {
+            // Record 的 getter 方法名与字段名相同
+            field.name
+        } else {
+            findGetterMethodName(psiClass, field)
+        }
+    }
+}
+
+private fun findGetterMethodName(psiClass: PsiClass, field: PsiField): String? {
+    // 1. 检查手动编写的 Getter 方法
+    val possibleNames = getPossibleGetterNames(psiClass, field)
+
+    return possibleNames.firstOrNull { name ->
+        psiClass.findMethodsByName(name, false).any { method ->
+            method.returnTypeElement?.type?.equals(field.type) == true
+        }
+    }
+}
+
+private fun getPossibleGetterNames(psiClass: PsiClass, field: PsiField): List<String> {
+    val fieldName = field.name
+    val isBoolean = field.type.canonicalText == "boolean" || field.type.canonicalText == "Boolean" || field.type.canonicalText == "java.lang.Boolean"
+
+    // 检查是否有 @Accessors 注解
+    val accessors = psiClass.modifierList?.annotations?.firstOrNull {
+        it.resolveAnnotationType()?.qualifiedName == "lombok.experimental.Accessors"
+    }
+
+    val names = mutableListOf<String>()
+
+    if (accessors != null) {
+        // 检查 fluent 模式
+        val fluent = accessors.findAttributeValue("fluent") as? PsiLiteralExpression
+        if (fluent?.value == true) {
+            // fluent 模式：直接使用字段名
+            names.add(fieldName)
+        } else {
+            // 检查 prefix 配置
+            val prefixAttr = accessors.findAttributeValue("prefix")
+            if (prefixAttr is PsiArrayInitializerMemberValue) {
+                prefixAttr.initializers.forEach { init ->
+                    if (init is PsiLiteralExpression) {
+                        val prefix = init.value?.toString() ?: return@forEach
+                        if (fieldName.startsWith(prefix)) {
+                            val baseName = fieldName.removePrefix(prefix)
+                            names.add(if (isBoolean) "is${baseName.replaceFirstChar { it.uppercase() }}"
+                            else "get${baseName.replaceFirstChar { it.uppercase() }}")
+                        }
+                    }
+                }
+            }
+        }
+    }
+
+    // 如果没有找到 Accessors 配置，使用标准命名
+    if (names.isEmpty()) {
+        val baseName = fieldName
+        names.add(if (isBoolean) "is${baseName.replaceFirstChar { it.uppercase() }}"
+        else "get${baseName.replaceFirstChar { it.uppercase() }}")
+    }
+
+    return names
+}
+
+fun getGetter(className: String, field: PsiField, map: Map<PsiField, String?>): String {
+    return if (map.containsKey(field) && map[field] != null) "$className::${map[field]}" else "o -> o.${field.name}"
+}
